Add the "install-task" mode
This mode provides a way to register with the Windows Task Scheduler
service so that TransitionFixer will start when the user logs into
Windows so that it can apply the proper transition fix.

In addition, change "install" to "install-log" (since the old "install"
command was used to register the program with the Windows Event Log so
that the user can see events for this program).